requirements.txt
google-generativeai>=0.3.0
python-bidi>=0.4.2

